+++ pkg/controller/controller.go
@@ -30,6 +30,7 @@ import (
 	"k8s.io/client-go/tools/leaderelection"
 	"k8s.io/klog/v2"
 
+	"github.com/piraeusdatastore/linstor-affinity-controller/pkg/util"
 	"github.com/piraeusdatastore/linstor-affinity-controller/pkg/version"
 )
 
@@ -58,11 +59,15 @@ type AffinityReconciler struct {
 }
 
 func NewReconciler(cfg *Config) (*AffinityReconciler, error) {
-	lclient, err := hlclient.NewHighLevelClient(client.Log(KLogV(3)))
+	lclient, err := hlclient.NewHighLevelClient(
+		client.Log(KLogV(3)),
+		client.UserAgent("linstor-affinity-controller/"+version.Version),
+	)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize linstor client: %w", err)
 	}
 
+	lclient.Resources = util.NewResourceCache(lclient.Resources, cfg.Timeout)
 	lclient.PropertyNamespace = cfg.PropertyNamespace
 
 	kclient, err := kubernetes.NewForConfig(cfg.RestCfg)
@@ -252,7 +257,7 @@ func (a *AffinityReconciler) getRemoteAccessParameter(ctx context.Context, pv *c
 	}
 
 	klog.V(3).Infof("Parsing storage class parameters '%s'", sc.Name)
-	params, err := volume.NewParameters(sc.Parameters)
+	params, err := volume.NewParameters(sc.Parameters, a.lclient.PropertyNamespace)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get storage class parameters: %w", err)
 	}